backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
pydantic==2.4.2
pydantic-core==2.10.1
pdfplumber==0.10.3
sentence-transformers==2.2.2
chromadb==0.4.15
python-dotenv==1.0.0
requests==2.31.0
uv==0.1.3
